backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
PyYAML==6.0.3
requests==2.20.0
fabric-sdk-py==1.0.0
grpcio==1.78.0
protobuf==3.19.6
cryptography==46.0.5
pycryptodomex==3.23.0
python-dotenv>=1.0.0
gunicorn>=23.0.0
ipfshttpclient==0.7.0
Werkzeug==3.1.7
itsdangerous==2.2.0
click==8.3.1
Jinja2==3.1.6
MarkupSafe==3.0.3
